+++ examples/privy/src/providers/SyncedWalletProvider.tsx
@@ -5,11 +5,12 @@ import {
 } from '@lifi/wallet-management'
 import { useAvailableChains } from '@lifi/widget'
 import { PrivyProvider } from '@privy-io/react-auth'
+import { SmartWalletsProvider } from '@privy-io/react-auth/smart-wallets'
 import { WagmiProvider } from '@privy-io/wagmi'
 import { QueryClientProvider } from '@tanstack/react-query'
 import type { FC, PropsWithChildren } from 'react'
 import { type Chain, mainnet } from 'viem/chains'
-import { PRIVY_APP_ID, PRIVY_CLIENT_ID, privyConfig } from '../config/privy'
+import { PRIVY_APP_ID, privyConfig } from '../config/privy'
 import { queryClient } from '../config/queryClient'
 import { wagmiConfig } from '../config/wagmi'
 import { SolanaProvider } from './SolanaProvider'
@@ -26,20 +27,22 @@ export const WalletProvider: FC<PropsWithChildren> = ({ children }) => {
   return (
     <PrivyProvider
       appId={PRIVY_APP_ID}
-      clientId={PRIVY_CLIENT_ID}
+      // clientId={PRIVY_CLIENT_ID}
       config={{
         ...privyConfig,
         defaultChain: mainnet,
         supportedChains,
       }}
     >
-      <QueryClientProvider client={queryClient}>
-        <SolanaProvider>
-          <WagmiProvider config={wagmiConfig} reconnectOnMount={false}>
-            {children}
-          </WagmiProvider>
-        </SolanaProvider>
-      </QueryClientProvider>
+      <SmartWalletsProvider>
+        <QueryClientProvider client={queryClient}>
+          <SolanaProvider>
+            <WagmiProvider config={wagmiConfig} reconnectOnMount={false}>
+              {children}
+            </WagmiProvider>
+          </SolanaProvider>
+        </QueryClientProvider>
+      </SmartWalletsProvider>
     </PrivyProvider>
   )
 }