+++ packages/widget/src/components/Chains/ChainList.tsx
@@ -2,6 +2,7 @@ import type { ExtendedChain } from '@lifi/sdk'
 import { Skeleton } from '@mui/material'
 import type { RefObject } from 'react'
 import { useTranslation } from 'react-i18next'
+import type { FormType } from '../../stores/form/types.js'
 import { SearchNotFound } from '../Search/SearchNotFound.js'
 import {
   List,
@@ -13,6 +14,7 @@ import { VirtualizedChainList } from './VirtualizedChainList.js'
 
 interface ChainListProps {
   parentRef: RefObject<HTMLDivElement | null>
+  formType: FormType
   chains: ExtendedChain[]
   onSelect: (chain: ExtendedChain) => void
   selectedChainId?: number
@@ -23,6 +25,7 @@ interface ChainListProps {
 
 export const ChainList = ({
   parentRef,
+  formType,
   chains,
   hasSearchQuery,
   onSelect,
@@ -79,6 +82,7 @@ export const ChainList = ({
   return (
     <VirtualizedChainList
       scrollElementRef={parentRef}
+      formType={formType}
       chains={chains}
       hasSearchQuery={hasSearchQuery}
       onSelect={onSelect}